refactor(news): extract BIP-322 signing to shared utils/bip322.ts
- Create src/utils/bip322.ts with three exported primitives:
  bip322TaggedHash, bip322BuildToSpendTxId, bip322Sign
- Update news.tools.ts to import bip322Sign from the shared util,
  removing ~95 lines of duplicated BIP-322 implementation
- Update signing.tools.ts to import bip322BuildToSpendTxId and
  bip322Sign from the shared util, removing the local duplicates
- Fix double getAccount() call in news_file_signal: buildNewsAuthHeaders
  now accepts a pre-fetched AccountForAuth instead of calling getAccount()
  internally, saving one async round trip per signal filed
- Remove unused RawTx import from signing.tools.ts
